Moved the stream cover view adapter to separate class, named StreamGridViewAdapter
