+++ app/src/main/java/com/example/apt_miniproject_android/ViewStreamsActivity.java
@@ -1,19 +1,13 @@
 package com.example.apt_miniproject_android;
 
-import android.content.Context;
 import android.content.Intent;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 import android.view.View;
-import android.view.ViewGroup;
 import android.view.WindowManager;
 import android.widget.AdapterView;
-import android.widget.BaseAdapter;
 import android.widget.Button;
 import android.widget.GridView;
-import android.widget.ImageView;
-import android.widget.LinearLayout;
-import android.widget.TextView;
 import android.widget.Toast;
 
 import com.example.apt_miniproject_android.backend.ServerCommunicator;
@@ -21,15 +15,11 @@ import com.example.apt_miniproject_android.backend.ServerResponseAction;
 import com.example.apt_miniproject_android.model.StreamInfo;
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
-import com.squareup.picasso.Picasso;
-
-import java.util.ArrayList;
-import java.util.List;
 
 public class ViewStreamsActivity extends AppCompatActivity {
 
     private GridView gridview;
-    private StreamCoverImageAdapter adapter;
+    private StreamGridViewAdapter adapter;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -40,7 +30,7 @@ public class ViewStreamsActivity extends AppCompatActivity {
 
         // set gridview adapter and click behavior
         gridview = (GridView) findViewById(R.id.gridview);
-        adapter = new StreamCoverImageAdapter(this);
+        adapter = new StreamGridViewAdapter(this);
         gridview.setAdapter(adapter);
 
         gridview.setOnItemClickListener(new AdapterView.OnItemClickListener() {
@@ -107,95 +97,4 @@ public class ViewStreamsActivity extends AppCompatActivity {
             }
         });
     }
-
-
-    public class StreamCoverImageAdapter extends BaseAdapter {
-        private Context mContext;
-        private List<StreamInfo> mstreamInfos;
-        private int width;
-        private GridView.LayoutParams parms;
-
-        public StreamCoverImageAdapter(Context c) {
-            mContext = c;
-            mstreamInfos = new ArrayList<>();
-        }
-
-        private void setWidth(int parent_width) {
-            int spacing = parent_width / 25;
-            width = (parent_width - 3 * spacing) / 4;
-
-            parms = new GridView.LayoutParams(width, 100 + width);
-            gridview.setHorizontalSpacing(spacing);
-            gridview.setVerticalSpacing(spacing);
-        }
-
-
-        public int getCount() {
-            return mstreamInfos.size();
-        }
-
-        public StreamInfo getItem(int position) {
-            return mstreamInfos.get(position);
-        }
-
-        public long getItemId(int position) {
-            return 0;
-        }
-
-        // create a new ImageView for each item referenced by the Adapter
-        public View getView(int position, View convertView, ViewGroup parent) {
-            myImageView imageView;
-
-            if (width != parent.getMeasuredWidth())
-                setWidth(parent.getMeasuredWidth());
-
-            imageView = new myImageView(mContext);
-            imageView.setLayoutParams(parms);
-
-            if(!mstreamInfos.get(position).getName().equals(""))
-                imageView.setStreamInfo(mstreamInfos.get(position), width);
-
-            return imageView;
-        }
-
-        public void addStreamInfo(StreamInfo info){
-            mstreamInfos.add(info);
-        }
-
-        private class myImageView extends LinearLayout {
-
-            private ImageView imageView;
-            private TextView textView;
-
-            public myImageView(Context context) {
-                super(context);
-                this.setOrientation(LinearLayout.VERTICAL);
-            }
-
-            public void setStreamInfo(final StreamInfo streamInfo, int width){
-
-                imageView = new ImageView(mContext);
-                imageView.setScaleType(ImageView.ScaleType.CENTER_CROP);
-                imageView.setLayoutParams(new GridView.LayoutParams(width, width));
-
-                textView = new TextView(mContext);
-
-                if(!streamInfo.getCoverImageURL().equals(""))
-                    Picasso.with(mContext)
-                            .load(streamInfo.getCoverImageURL())
-                            .placeholder(android.R.drawable.picture_frame)
-                            .into(imageView);
-
-                textView.setText(streamInfo.getName());
-
-                this.addView(imageView);
-                this.addView(textView);
-            }
-
-
-        }
-
-    }
-
-
 }
